Rename file, improve token generator logging
