+++ tests/k6/tests/serviceowner/all-tests.js
@@ -1,12 +1,12 @@
 // This file is generated, see "scripts" directory
 import { default as authentication } from './authentication.js';
-import { default as dialogCreate } from './dialogCreate.js';
 import { default as dialogCreateInvalidActionCount } from './dialogCreateInvalidActionCount.js';
+import { default as dialogCreatePatchDelete } from './dialogCreatePatchDelete.js';
 import { default as dialogSearch } from './dialogSearch.js';
 
 export default function() {
   authentication();
-  dialogCreate();
   dialogCreateInvalidActionCount();
+  dialogCreatePatchDelete();
   dialogSearch();
 }
